- move contract module&service to common; - fix historical aggregation; - fix imports;
