+++ apps/aggregator/src/aggregator.module.ts
@@ -1,4 +1,5 @@
 import { ClassSerializerInterceptor, Module } from '@nestjs/common';
+import { APP_INTERCEPTOR } from '@nestjs/core';
 import { ConfigModule } from '@nestjs/config';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { ScheduleModule } from '@nestjs/schedule';
@@ -13,16 +14,15 @@ import { RedisModule } from '@dao-stats/redis';
 import { TransactionModule } from '@dao-stats/transaction';
 import { HttpCacheModule } from '@dao-stats/cache';
 import {
+  CoinPriceHistoryModule,
+  ContractModule,
   DaoStatsModule,
   DaoStatsHistoryModule,
   DaoModule,
 } from '@dao-stats/common';
-import { AggregatorService } from './aggregator.service';
-import { APP_INTERCEPTOR } from '@nestjs/core';
 import { ReceiptModule } from '@dao-stats/receipt';
-import { CoinPriceHistoryModule } from '@dao-stats/common/coin-price-history.module';
-import { ContractModule } from 'apps/api/src/contract/contract.module';
-import { ExchangeModule } from 'libs/exchange/src/exchange.module';
+import { ExchangeModule } from '@dao-stats/exchange';
+import { AggregatorService } from './aggregator.service';
 
 @Module({
   imports: [
