+++ apps/api/src/contract/contract.module.ts
@@ -1,21 +1,19 @@
-import { TypeOrmModule } from '@nestjs/typeorm';
 import { CacheModule, Module } from '@nestjs/common';
 
 import { CacheConfigService } from '@dao-stats/config/cache';
-import { Contract } from '@dao-stats/common';
+import { ContractModule, ContractService } from '@dao-stats/common';
 
 import { ContractController } from './contract.controller';
-import { ContractService } from './contract.service';
 
 @Module({
   imports: [
     CacheModule.registerAsync({
       useClass: CacheConfigService,
     }),
-    TypeOrmModule.forFeature([Contract]),
+    ContractModule,
   ],
   providers: [ContractService],
   controllers: [ContractController],
   exports: [ContractService],
 })
-export class ContractModule {}
+export class ApiContractModule {}
