--- apps/api/src/contract/dto/index.ts
@@ -1,1 +1,0 @@
-export * from './contract.dto';
