+++ libs/exchange/src/exchange.module.ts
@@ -2,7 +2,8 @@ import { Module } from '@nestjs/common';
 import { HttpModule } from '@nestjs/axios';
 import { ConfigModule } from '@nestjs/config';
 
-import { CoinGeckoService, SodakiService } from '.';
+import { CoinGeckoService } from './coin-gecko.service';
+import { SodakiService } from './sodaki.service';
 
 @Module({
   imports: [ConfigModule, HttpModule],
