+++ libs/exchange/src/index.ts
@@ -1,2 +1,4 @@
-export * from './sodaki.service';
+export * from './types';
 export * from './coin-gecko.service';
+export * from './exchange.module';
+export * from './sodaki.service';
