+++ libs/exchange/src/sodaki.service.ts
@@ -3,10 +3,9 @@ import { HttpService } from '@nestjs/axios';
 import { Injectable } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { Cacheable } from '@type-cacheable/core';
-import { CoinType } from '@dao-stats/common/types/coin-type';
-import { CurrencyType } from '@dao-stats/common/types/currency-type';
-import { SodakiCoinTypeSymbol } from './types/sodaki-coin-type-symbol';
+import { CoinType, CurrencyType } from '@dao-stats/common';
 import { ExchangeConfig } from '@dao-stats/config/exchange-config';
+import { SodakiCoinTypeSymbol } from './types';
 
 @Injectable()
 export class SodakiService {
